requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0
drf-yasg>=1.21.7
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
